requirements.txt
Django==2.2.7
django-cors-headers==3.2.0
django-filter==2.2.0
djangorestframework==3.10.3
Markdown==3.1.1
pytz==2019.3
sqlparse==0.3.0
gunicorn
